[ECR-2116] Adapter classes renamed.
